# More aggressive ahead of time optimization for string-key type formatter.