… src/Utf8Json.CodeGenerator/Generator/FormatterTemplate1.cs
… src/Utf8Json.CodeGenerator/Generator/StringKey/ULongGroupEnumerable.cs
… src/Utf8Json.CodeGenerator/Utf8Json.CodeGenerator.csproj
… src/Utf8Json.CodeGenerator/packages.config
… src/Utf8Json.UniversalCodeGenerator/Generator/FormatterTemplate1.cs
… src/Utf8Json.UniversalCodeGenerator/Utf8Json.UniversalCodeGenerator.csproj